Send and receive with packet header

+++ util.hpp
@@ -9,7 +9,7 @@ uint32_t buf2int(const char *s, size_t a, size_t b) {
 }
 
 void int2buf(char *buf, uint32_t num, size_t a, size_t b) {
-    for (int i=b-1; i>=a; i--) {
+    for (int i=b-1; i>=(int) a; i--) {
         buf[i] = num % 256;
         num /= 256;
     }
